changed parameter int to double

+++ Edge.java
@@ -3,7 +3,7 @@ public class Edge {
     private Vertex destination;
     private double weight;
 
-    public Edge(Vertex source, Vertex destination, int weight) {
+    public Edge(Vertex source, Vertex destination, double weight) {
         this.source = source;
         this.destination = destination;
         this.weight = weight;
